+++ mrblib/acli/utils.rb
@@ -1,7 +1,7 @@
 module Acli
   module Utils
     def exit_with_error(e, code = 1)
-      puts "Error: #{e.message}"
+      puts "Error: #{e.message}\n#{e.backtrace.join("\n")}"
       exit code
     end
 
@@ -27,6 +27,24 @@ module Acli
       else
         str
       end
+    end
+
+    # Downcase and prepend with protocol if missing
+    def normalize_url(url)
+      url = url.downcase
+      # Replace ws protocol with http, 'cause URI cannot resolve port for non-HTTP
+      url.sub!("ws", "http")
+      url = "http://#{url}" unless url.start_with?("http")
+      url
+    end
+
+    def uri_to_ws_s(uri)
+      "#{(uri.scheme == "https" || uri.scheme == "wss") ? "wss://" : "ws://"}"\
+      "#{uri.userinfo ? "#{uri.userinfo}@" : ""}"\
+      "#{uri.host}"\
+      "#{uri.port == 80 || (uri.port == 443 && uri.scheme == "https") ? "" : ":#{uri.port}"}"\
+      "#{uri.path}"\
+      "#{uri.query ? "?#{uri.query}" : ""}"
     end
 
     extend self